# src/backend/src/ApplicationManagerAPI/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Application Manager API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/FollowUp": {
      "get": {
        "tags": [
          "FollowUp"
        ],
        "operationId": "FollowUp_GetAll",
        "parameters": [
          {
            "name": "jobApplicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowUpDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FollowUp"
        ],
        "operationId": "FollowUp_Create",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowUpRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/FollowUp/{id}": {
      "put": {
        "tags": [
          "FollowUp"
        ],
        "operationId": "FollowUp_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFollowUpRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FollowUp"
        ],
        "operationId": "FollowUp_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "FollowUpDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "guid"
          },
          "dueDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          }
        }
      },
      "CreateFollowUpRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobApplicationId"
        ],
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "dueDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          }
        }
      },
      "UpdateFollowUpRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobApplicationId"
        ],
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "dueDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}